requirements.txt
aiosqlite==0.17.0
click==7.1.2
h11==0.12.0
iso8601==0.1.14
pip==21.0.1
PyPika==0.47.7
pytz==2021.1
setuptools==54.1.1
starlette==0.14.2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.4
